backend/requirements-dev.txt
-r requirements.txt
ruff==0.4.4
mypy==1.10.0
django-stubs[compatible-mypy]==5.0.0
djangorestframework-stubs==3.15.0
pytest==8.2.2
pytest-django==4.8.0
types-requests==2.32.0.20240521
types-psycopg2==2.9.21.20240311
